fix(preview.2): address #1344 CodeRabbit review batch (12/15)
Code (5 files):

- workflow-status-view: when `_taskStore` is missing on a pre-2A.7
  snapshot, backfill from the legacy `_seenAssignedTaskIds` /
  `_seenCompletedTaskIds` maps + `tasksFailed` aggregate. `task.*` event
  fold now keeps working across the upgrade boundary instead of
  crashing (`undefined` → `state.tasks`) or silently dropping the
  surface counts.

- event-store/decide.test.ts: extract `makeFixtureReducer`, `seedStream`,
  `FixtureState` into new `decide-fixtures.ts`. Three sibling tests
  (aggregate-stream, with-session, decide.race) imported these from a
  `*.test.ts` module, which registers all of `decide.test.ts`'s `describe`
  blocks as side effects on import.

- event-store/grep-gates.test.ts: derive `__dirname` from `import.meta.url`
  — the bare reference fails at module load under NodeNext ESM.

Skills (5 source files, 36 generated variants):

- skills-src/workflow-state/references/mcp-tool-reference.md: replace
  the stale `set`-only action list with the v2.11 surface
  (init/get/transition/update/cancel/cleanup/reconcile/rehydrate/
  checkpoint/describe); split the action table into `transition` and
  `update` rows; rewrite GUARD_FAILED guidance for the two-call pattern.

- skills-src/workflow-state/references/phase-transitions.md: rewrite
  the "Combined updates + phase" pattern as
  "apply-updates-then-transition" — `update` rejects `phase` in its
  payload (registry.ts:1003), so the single-call combo no longer works;
  the structured `INVALID_INPUT` + `suggestedFix` is exactly the split
  this doc now documents.

- skills-src/{quality-review,delegation}: replace `action: "update", …,
  phase: "X"` examples with `action: "transition", …, target: "X"` in
  auto-transition.md, SKILL.md, and state-management.md.

Deferred (3 comments on files wave-a is concurrently editing): two on
storage/sqlite-backend.ts and one on projections/store.ts — addressing
those on release would conflict with wave-a's `resolveMaxRecords`
extraction. Will fold in after wave-a merges to release.
