+++ servers/exarchos-mcp/src/event-store/with-session.test.ts
@@ -13,7 +13,7 @@ import {
   createRegistry,
   type ProjectionRegistry,
 } from '../projections/registry.js';
-import { makeFixtureReducer, seedStream, type FixtureState } from './decide.test.js';
+import { makeFixtureReducer, seedStream, type FixtureState } from './decide-fixtures.js';
 
 /**
  * Wave 3 Tasks 3.8 – 3.10 — `withSession<TState>` imperative-escape-hatch
