+++ skills/cursor/workflow-state/references/mcp-tool-reference.md
@@ -10,7 +10,7 @@ Unified MCP server for workflow orchestration, event sourcing, CQRS views, and t
 
 | Tool | Actions | When to Use |
 |------|---------|-------------|
-| `mcp__plugin_exarchos_exarchos__exarchos_workflow` | `init`, `get`, `set`, `cancel`, `cleanup` | Workflow CRUD: starting workflows, reading/updating state, cancelling abandoned workflows, resolving merged workflows |
+| `mcp__plugin_exarchos_exarchos__exarchos_workflow` | `init`, `get`, `transition`, `update`, `cancel`, `cleanup`, `reconcile`, `rehydrate`, `checkpoint`, `describe` | Workflow CRUD: starting workflows, reading/updating state, transitioning phases, cancelling abandoned workflows, resolving merged workflows |
 | `mcp__plugin_exarchos_exarchos__exarchos_event` | `append`, `query` | Event sourcing: recording workflow events, reading event history |
 | `mcp__plugin_exarchos_exarchos__exarchos_orchestrate` | `task_claim`, `task_complete`, `task_fail` | Task coordination and lifecycle |
 | `mcp__plugin_exarchos_exarchos__exarchos_view` | `pipeline`, `tasks`, `workflow_status`, `stack_status`, `stack_place` | CQRS materialized views for read-optimized queries |
@@ -22,7 +22,8 @@ Unified MCP server for workflow orchestration, event sourcing, CQRS views, and t
 |--------|-------------|
 | `init` | Starting any `/exarchos:ideate`, `/exarchos:debug`, or `/exarchos:refactor` workflow |
 | `get` | Restoring context, checking phase, reading task details. Use `query` for dot-path lookup (e.g., `query: "phase"`), or `fields` array for projection (e.g., `fields: ["phase", "tasks"]`) to reduce token cost |
-| `set` | Updating phase (`phase: "delegate"`), recording artifacts, marking tasks complete. Use `updates` for field changes and `phase` for transitions |
+| `transition` | Advance to a target phase (`target: "delegate"`). Canonical phase-mutation surface — emits `workflow.transition`, runs HSM guards, returns `validTargets` + `expectedShape` on guard failure |
+| `update` | Mutate non-phase state fields (`updates: { "artifacts.design": "..." }`). Rejects `phase` in `updates` with a structured `INVALID_INPUT` pointing at `transition`. Emits `state.patched` |
 | `cancel` | Cleaning up abandoned workflows. Supports `dryRun: true` to preview cleanup actions |
 | `cleanup` | Resolve a merged workflow to completed. Verifies merge, backfills synthesis metadata, force-resolves reviews, transitions to completed. Requires `mergeVerified: true` — pass after verifying PRs are merged via GitHub API |
 
@@ -74,7 +75,7 @@ Unified MCP server for workflow orchestration, event sourcing, CQRS views, and t
 No path exists from current phase to target. Check `validTargets` in the error — it lists reachable phases with guard descriptions. You may need to step through intermediate phases.
 
 ### GUARD_FAILED
-The transition exists but the guard condition is unmet. Send prerequisite `updates` and `phase` in a **single** `set` call — updates apply before guards evaluate. See `references/phase-transitions.md` for guard prerequisites.
+The transition exists but the guard condition is unmet. Apply the prerequisite state with `update`, then retry the `transition` — the `update`/`transition` split is enforced by the workflow tool (phase changes go through HSM guards, non-phase mutations don't). See `references/phase-transitions.md` for guard prerequisites.
 
 ### CIRCUIT_OPEN
 A compound state's fix cycle limit was reached. Escalate to user or cancel the workflow.
